* fix(EventsPresenter.kt): fix screen navigation in EventsPresenterFactory * chore(EventsPagingSource.kt): add logging to EventsPagingSource The screen navigation in the EventsPresenterFactory was incorrect, causing the wrong screen to be displayed. This has been fixed. In the EventsPagingSource, a logging statement has been added to print the number of entries being loaded. This is to aid in debugging and monitoring the application.
* refactor(EventsRemoteMediator.kt): replace EventsService with EventsApi and LegacyEvent
* chore(EventsRequest.kt): delete unused file
The EventsRemoteMediator class now uses the EventsApi instead of the EventsService. The LegacyEvent class is now used to wrap the response from the EventsApi. The EventsRequest.kt file is no longer used and has been deleted.

* refactor(EventsService.kt): replace custom HTTP request with Ktor HTTP client
* chore(HomeScreen.kt): rename view model to HomeViewModel
The custom HTTP request in EventsService.kt has been replaced with Ktor HTTP client, which simplifies the code and improves readability. In HomeScreen.kt, the view model has been renamed to HomeViewModel to better reflect its purpose.

* refactor(HomeViewModel.kt): rename EventsHomeViewModel to HomeViewModel and update constructor parameter name
* feat(HomeViewModel.kt): use remember to create EventsApi instance
The class name has been updated to improve consistency with the naming conventions. The constructor parameter name has been updated to improve readability. The remember function is now used to create an instance of the EventsApi class, which improves performance by ensuring that the same instance is used across recompositions.

* feat(ApiBuilder.kt): add buildApi function to build an API client with an API key
* chore(HttpClient.kt): import HttpHeaders from ktor.http
* feat(LegacyEvent.kt): add LegacyEvent function to convert Event to LegacyEvent
The buildApi function allows for the creation of an API client with an API key. This is useful for authenticating requests to the API. The HttpHeaders import is added to HttpClient.kt to allow for the use of HttpHeaders in the code. The LegacyEvent function is added to convert an Event object to a LegacyEvent object. This is useful for backwards compatibility with older versions of the API.

* fix(Events.sq): change selectAll query to order by dateStart in descending order
The selectAll query was modified to order by dateStart in descending order instead of ascending order. The WHERE clause was also changed to filter events that occur before or at the specified startAt date.