* refactor(EventsPresenter.kt): rename Event class to LegacyEvent * feat(EventsPresenter.kt): add Details data class to EventsScreen sealed interface * feat(EventsPresenter.kt): import DetailsScreen and onNavEvent The Event class has been renamed to LegacyEvent to avoid naming conflicts with other classes. A new data class called Details has been added to the EventsScreen sealed interface to represent the details screen. The DetailsScreen and onNavEvent have been imported to support the new Details data class.
* refactor(EventsPresenter.kt): add NavEvent to EventsEvent sealed interface and refactor EventsPresenter to handle NavEvent
* feat(EventsPresenter.kt): add DetailsScreen to EventsUiFactory
The NavEvent is added to the EventsEvent sealed interface to allow for navigation to a specific screen. The EventsPresenter is refactored to handle the NavEvent and navigate to the specified screen. The DetailsScreen is added to the EventsUiFactory to allow for displaying details of an event.

* chore(DetailsScreen.kt): add DetailsScreen composable function
* refactor(EventsPagingSource.kt): remove println statement
The DetailsScreen composable function is added to the project. It takes in the current state and a modifier as parameters and displays the name of the current event. The println statement in the EventsPagingSource.kt file is removed as it is no longer needed.

* refactor(HomeScreen.kt): change onClick function signature in EventSection
* feat(HomeScreen.kt): pass event to onClick function in EventSection
The onClick function signature in the EventSection composable has been changed to accept an Event object as a parameter. This allows the event object to be passed to the onClick function when the button is clicked. The onClick function is now called with the event object instead of just printing a message to the console. This improves the functionality of the app by allowing the user to navigate to the details screen for the selected event.